ui: use a different apollo client for /products UI
The products shema has a GQL type named Payment that conflicts with
the type named Payment in the payments schema. This causes the
introspection fragment matcher codegen to fail. So instead we wrap
the products app in its own client.

Note: a later commit should probably do the same thing for the other
apps.