ui: update final form typings and breaks
Instead of creating the decorators inside the generic Form in @pay/web
we now make the callers create the decorators. This seems necessary
because the Values generic type needs to be passed to createDecorator.
Doing this inside the render function of Form causes a warning to be
generated by Final Form telling us not to recreate these decorators. The
sensible thing seems to be to just make the caller create the
decorators.

Now use @types/final-form-focus instead of our own hand-written typings.